8012345678 is a better mobile example for Japan
The current JP mobile example number 7012345678, while technically valid, looks strange to most Japanese people. 070- mobile numbers are reserved for PHS (Personal Handy-phone System) which is a legacy technology that is not commonly used. 080- and 090- numbers are the most common. In general (though not 100% strict) 080- tend to be for smartphones and 090- are for legacy mobile phones, hence I've chosen 080.