Updating build process to generate new possible length data (#1308)
https://groups.google.com/forum/#!topic/libphonenumber-discuss/75TOpTFVi08

* Initial code changes to support the new possibleLengths infrastructure
when building metadata. Includes setting these in the main metadata
file and changing the code to populate them in the PhoneMetadata proto
at build time. Updating tests for Java.

* Updating comment about the possible_lengths_local_only field.

* Adding the generated jars that build files containing the new possible
lengths fields.

* Rebuilding short-number metadata with the new proto field. Not doing
this at release time so we can see what changes in the metadata are due
to metadata fixes, rather than this possibleLengths field.

* Regeneration of phone number metadata with the new possible length
information filled in.

Includes some possible length changes for KR and BY, and a validation
fix for NA where a digit was missing.

* Updating the test proto metadata files with the new possible length
information.

* Added notes about the code changes.

* Regenerating C++ metadata with new possibleLength info and updating unit
test to check this works.