Modified the pom.xml. Now the README instructions work (I had to do a mvn clean install from the java/ dir first). Re-ordered it to match the geocoder pom.xml file for easier diffing.
Actual diffs:
-- added two attributes the geocoder one had (url and packaging) - not
sure why we have them but should be consistent
-- reordered the "build" and "dependencies" and "parent" sections
-- deleted a section.

Tested: will be interesting to see if Travis likes it, but it at least
worked for me. That said, maybe it would have worked before I made the
changes if I had known to do "mvn clean install". So the changes might
not in fact be necessary - if they don't fix the Travis build, I will
revert (may keep the reordering for ease of use later).