Including local-only length data for subtypes in the build.
Using these to provide two more public methods: isPossibleNumberForType
and isPossibleNumberForTypeWithReason methods, that look at the number
lengths for a particular *type* and decide whether it is feasible or
not. If this is called for a type that is not supported for a region,
INVALID_LENGTH is returned.

Adding getSupportedTypesForRegion and getSupportedTypesForNonGeoEntity
too, so people can call isPossibleNumberForTypeWithReason *knowing* if
there even are numbers of this type for a country.

Updating the test metadata file and generated files.

This patch is Java only, C++ and JS coming.