Update Unleashed prefix
Starting in 2018, Mobile Vikings will also distribute new numbers with
prefix 456. Since Stievie is not a mobile brand, it can be ignored in
the listing.

As these mobile numbers are not yet distributed, there is no official
statement confirming this. The following FAQ page however already
mentions a 32456 number in Mobile Vikings context:
https://support.vikingco.com/hc/en-us/articles/360000313477-What-will-happen-after-activation-on-the-new-platform-