German Short Numbers of 118xx and 1180xx
The German Short Numbers 118xx and 1180xx as defined in https://www.bundesnetzagentur.de/SharedDocs/Downloads/DE/Sachgebiete/Telekommunikation/Unternehmen_Institutionen/Nummerierung/Rufnummern/118xy/118xyNummernplan.pdf?__blob=publicationFile&v=2 are added with this proposal. I do not know of [123456789]\d{1} would be a valid notation, so I wrote each ciffer in a separated line, because of the 0 as been a special case.