Use a system-installed abseil when available
Currently, CMake is configured to download and build abseil from source
whenever `BUILD_GEOCODER` is set to `ON`. Let's allow users to use a
pre-existing installation of abseil by checking for one first before
attempting to download and build another one.

If CMake finds a pre-existing installation of abseil, it will use that
instead.

We need to allow the user to override `CMAKE_CXX_STANDARD` because they
may have built abseil with a newer standard than C++11. Attempting to
build libphonenumber with C++11 will fail if abseil was built with a
newer C++ standard.

Closes #2772.