Make C++ standard handling more consistent
In 035901b78dbe5a6fddc9e758f0069c310750db72, we updated
`cpp/CMakeLists.txt` to respect the user's configuration of
`CMAKE_CXX_STANDARD`.

It turns out that this is not sufficient for some build configurations,
because the user's configuration is still overridden inside
`tools/cpp/CMakeLists.txt`.

Let's fix that so that the handling of `CMAKE_CXX_STANDARD` is
consistent throughout the project.