Update Danish number plan according to the SDFI
The information can be found via:
https://eng.sdfi.dk/digital-infrastructure/telecom/numbering

The link to the number plan is this:
* https://sdfi.dk/Media/637925268968845027/nummerplan_2020_farver.pdf

Notable exceptions are in landline segments where chunks are allocated
to mobile numbers:

- In number series marked with (!) the following number series are
  primarily allocated for mobile communication instead of landline
  communication: 342, 344-349, 356-357, 359, 362, 365-366, 389, 398,
  431, 441, 462, 466, 468, 472, 474, 476, 478, 485-486, 488-489,
  493-496, 498-499, 542-543, 545, 551-552, 556, 571-574, 577, 579, 584,
  586-587, 589, 597-598, 627, 629, 641, 649, 658, 662-665, 667, 692-694,
  697, 771-772, 782-783, 785-786, 788-789, 826-827 and 829.

These exceptions are listed as mobile numbers in the XML and are still
present in the fixed-line configuration.

The Danish should start to indicate when these allocations are
finalized.

Signed-off-by: Wesley Schwengle <wesleys@opperschaap.net>